+++ foundation_stdlib/init.lua
@@ -1,4 +1,4 @@
-local mod = foundation.new_module("foundation_stdlib", "1.16.0")
+local mod = foundation.new_module("foundation_stdlib", "1.17.0")
 
 mod:require("lib/encoding_tables.lua")